% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_ar_loglik}
\alias{cal_ar_loglik}
\title{Calculate log-likelihood for AR(p) model}
\usage{
cal_ar_loglik(X_list, Y_list, phi_hat, n_ts)
}
\arguments{
\item{X_list}{List of AR design matrices.}

\item{Y_list}{List of response vectors.}

\item{phi_hat}{Estimated AR coefficients (N x (p+1)).}

\item{n_ts}{Time series length.}
}
\value{
log-likelihood.
}
\description{
Calculate log-likelihood for AR(p) model
}
