% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{ClustAR1}
\alias{ClustAR1}
\title{Clustering nodes.}
\usage{
ClustAR1(
  TS_by_node,
  adj_w,
  ADMM_iter,
  GD_iter,
  lr,
  lambda,
  gamma,
  update_gamma,
  verbose
)
}
\arguments{
\item{TS_by_node}{Numeric matrix of size N x n. Each row is the time series for a node.}

\item{adj_w}{Adjacency matrix with non-negative weights}

\item{ADMM_iter}{ADMM iteration}

\item{lambda}{GFL penalty}

\item{gamma}{penalty for the augmentation term}

\item{update_gamma}{If TRUE, gamma is updated with a schedule}

\item{verbose}{If TRUE, print info during parameter learning}
}
\value{
A list
}
\description{
This function clusters nodes.
}
\examples{
N = 30 # nodes
n = 100 # time points
mu = 0
sigma = 1
phi = 0.8 # AR parameter
}
