% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{ClustARp}
\alias{ClustARp}
\title{Clustering nodes.}
\usage{
ClustARp(
  TS_by_node,
  adj_w,
  lag_p,
  ADMM_iter,
  lambda,
  gamma,
  update_gamma,
  verbose
)
}
\arguments{
\item{TS_by_node}{Numeric matrix of size N x n. Each row is the time series for a node.}

\item{adj_w}{Adjacency matrix with non-negative weights}

\item{ADMM_iter}{ADMM iteration}

\item{lambda}{GFL penalty}

\item{gamma}{penalty for the augmentation term}

\item{update_gamma}{If TRUE, gamma is updated with a schedule}

\item{verbose}{If TRUE, print info during parameter learning}
}
\value{
A list
}
\description{
This function clusters nodes.
}
\examples{
ts_data <- gen_ar1_by_cluster(cluster_sizes = c(50, 50, 50), n = 100)
graph_data <- gen_weighted_block_graph(cluster_sizes = c(50, 50, 50))
TS_by_node <- ts_data$TS_by_node
adj_w <- graph_data$weights
labels <- ts_data$labels
result <- ClustARp(TS_by_node, adj_w, lag_p=1, ADMM_iter=200, lambda=1, gamma=1,
                   update_gamma=FALSE, verbose=TRUE)
plot(result$phi, col=labels)

}
