% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{gen_ar1_by_cluster}
\alias{gen_ar1_by_cluster}
\title{Generate nodal AR(1) time series}
\usage{
gen_ar1_by_cluster(
  cluster_sizes = c(50, 50, 50),
  n = 100,
  mu_vec = c(0, 2, -2),
  phi_vec = c(0.2, 0.6, -0.4),
  sigma_vec = c(0.5, 0.5, 0.5),
  stationary_init = TRUE,
  seed = NULL
)
}
\description{
Generate nodal AR(1) time series
}
