% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ar_X_Y}
\alias{get_ar_X_Y}
\title{Build AR(p) design matrices for each node}
\usage{
get_ar_X_Y(TS_by_node, lag_p)
}
\arguments{
\item{TS_by_node}{Numeric matrix of size N x n. Each row is the time series for a node.}

\item{lag_p}{Integer lag p for the AR(p) model.}
}
\value{
A list with two elements:
\describe{
\item{X_list}{List of length N; each X is (n - p) x (p + 1) design matrix for node i.}
\item{Y_list}{List of length N; each Y is length (n - p) response vector for node i.}
}
}
\description{
Build AR(p) design matrices for each node
}
