% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_graph_info}
\alias{get_graph_info}
\title{Get graph information}
\usage{
get_graph_info(W)
}
\arguments{
\item{W}{Numeric matrix (N x N) weighted adjacency matrix. Must be symmetric.}
}
\value{
A list with:
\itemize{
\item edge_list: numeric matrix (i, j, w_ij)
\item node_degree: integer vector of node degrees
}
}
\description{
Convert weighted adjacency matrix to (1) edge list and (2) node degrees
}
\examples{
# 5-node weighted, symmetric adjacency matrix
W <- matrix(0, nrow = 5, ncol = 5)
W[1, 2] <- 1.5; W[1, 4] <- 2.0; W[2, 3] <- 0.8; W[4, 5] <- 3.2
W <- W + t(W)
get_graph_info(W)
}
