% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GraphClust_AR1_cpp_deprecate}
\alias{GraphClust_AR1_cpp_deprecate}
\title{ADMM for GraphClustAR}
\usage{
GraphClust_AR1_cpp_deprecate(
  X_list,
  Y_list,
  edge_list,
  node_degree,
  ADMM_iter,
  T_length,
  lambda,
  gamma,
  update_gamma,
  verbose
)
}
\arguments{
\item{X_list}{List of length N; each \code{X_list[[i]]} is (n-p x p) design matrix X_i.}

\item{Y_list}{List of length N; each \code{Y_list[[i]]} is length n-p response vector Y_i.}

\item{edge_list}{edge list (E by 3)}

\item{node_degree}{Numeric vector of length N. This is |B(i)| for each i.}

\item{ADMM_iter}{ADMM iteration}

\item{T_length}{length of nodal time series}

\item{lambda}{GFL penalty parameter.}

\item{gamma}{penalty for the augmentation term.}

\item{update_gamma}{If TRUE, gamma is updated with a schedule}

\item{verbose}{If TRUE, print info during learning.}
}
\value{
An N x d matrix with the updated phi.
}
\description{
ADMM for GraphClustAR (cpp version)
This function will not be export in the final version.
}
