% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{GraphClustARp}
\alias{GraphClustARp}
\title{Clustering nodes.}
\usage{
GraphClustARp(
  TS_by_node,
  adj_w,
  lag_p,
  num_clust_list,
  lambda_list,
  ADMM_iter,
  phi_tol = 0.001,
  verbose = F,
  update_gamma = F
)
}
\arguments{
\item{TS_by_node}{Numeric matrix of size N x n. Each row is the time series for a node.}

\item{adj_w}{Adjacency matrix with non-negative weights}

\item{num_clust_list}{vector containing list of cluster number}

\item{lambda_list}{vector containing list of GFL penalty}

\item{ADMM_iter}{ADMM iteration}

\item{phi_tol}{ADMM stopping criteria}

\item{verbose}{If TRUE, print info during parameter learning}

\item{update_gamma}{If TRUE, gamma is updated with a schedule}
}
\value{
A list
}
\description{
This function clusters nodes.
}
\examples{
ts_data <- gen_ar1_by_cluster(cluster_sizes = c(50, 50, 50), n = 100, seed=123)
graph_data <- gen_weighted_block_graph(cluster_sizes = c(50, 50, 50), seed=123)
TS_by_node <- ts_data$TS_by_node
adj_w <- graph_data$adj_w
result <- GraphClustARp(TS_by_node, adj_w, lag_p=1, num_clust_list=2:7,
                   lambda_list=c(0.1,0.25,0.5,0.75,1), ADMM_iter=100)
result$lambda
result$K
result$cluster
plot(result$ADMM_output$phi, col=result$cluster)

}
