% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_eta}
\alias{update_eta}
\title{Update theta}
\usage{
update_eta(theta, nu, eta, edge_list)
}
\arguments{
\item{theta}{E by d matrix of scaled dual variables.}

\item{nu}{E by d matrix of auxiliary variables.}

\item{edge_list}{E by 3 matrix. Each row e is (i, j, w_ij)}

\item{phi}{N by d matrix of node parameters.}
}
\value{
Updated nu matrix of size E by p.
}
\description{
ADMM update for theta
This function will not be export in the final version.
}
