% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_mu_AR1}
\alias{update_mu_AR1}
\title{Update mu of AR(1)}
\usage{
update_mu_AR1(
  X_list,
  Y_list,
  phi,
  mu,
  nu,
  theta,
  edge_list,
  node_degree,
  gamma,
  T_length
)
}
\arguments{
\item{X_list}{List of length N; each \code{X_list[[i]]} is (n-p x d) design matrix X_i.}

\item{Y_list}{List of length N; each \code{Y_list[[i]]} is length n-p response vector Y_i.}

\item{phi}{N vector of node parameters.}

\item{nu}{Edge-wise auxiliary variables, E vector.}

\item{theta}{Edge-wise scaled dual variables, E vector.}

\item{edge_list}{edge list (E by 3)}

\item{node_degree}{Numeric vector of length N. This is |B(i)| for each i.}

\item{gamma}{penalty for the augmentation term}

\item{T_length}{length of nodal time series}
}
\value{
An N vector with the updated mu.
}
\description{
ADMM update for mu using edge-wise neighbor aggregation.
This function will not be export in the final version.
}
