% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_nu}
\alias{update_nu}
\title{Update nu}
\usage{
update_nu(phi, theta, edge_list, lambda, gamma)
}
\arguments{
\item{phi}{N x p matrix of node parameters (one row per node).}

\item{theta}{E x p matrix of scaled dual variables.}

\item{edge_list}{E x 3 matrix. Each row e is (i, j, w_ij)}

\item{lambda}{GFL penalty parameter.}

\item{gamma}{penalty parameter for augmentation term.}
}
\value{
Updated nu matrix of size E x p.
}
\description{
ADMM update for nu using group soft-thresholding.
This function will not be export in the final version.
}
