% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_nu_AR1}
\alias{update_nu_AR1}
\title{Update nu of AR(1)}
\usage{
update_nu_AR1(mu, theta, edge_list, lambda, gamma)
}
\arguments{
\item{mu}{N vector of node parameters.}

\item{theta}{Edge-wise scaled dual variables, E vector.}

\item{edge_list}{edge list (E by 3)}

\item{lambda}{penalty parameter}

\item{gamma}{penalty for the augmentation term}

\item{node_degree}{Numeric vector of length N. This is |B(i)| for each i.}
}
\value{
An N vector with the updated mu.
}
\description{
ADMM update for mu using edge-wise neighbor aggregation.
This function will not be export in the final version.
}
