% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_phi}
\alias{update_phi}
\title{Update phi}
\usage{
update_phi(X_list, Y_list, phi, nu, theta, edge_list, node_degree, gamma)
}
\arguments{
\item{X_list}{List of length N; each \code{X_list[[i]]} is (n-p x d) design matrix X_i.}

\item{Y_list}{List of length N; each \code{Y_list[[i]]} is length n-p response vector Y_i.}

\item{phi}{N x d matrix of node parameters.}

\item{nu}{Edge-wise auxiliary variables, E x d matrix.}

\item{theta}{Edge-wise scaled dual variables, E x d matrix.}

\item{edge_list}{edge list (E by 3)}

\item{node_degree}{Numeric vector of length N. This is |B(i)| for each i.}

\item{gamma}{penalty for the augmentation term}
}
\value{
An N x d matrix with the updated phi.
}
\description{
ADMM update for phi using edge-wise neighbor aggregation.
This function will not be export in the final version.
}
