% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_phi_AR1}
\alias{update_phi_AR1}
\title{Update phi of AR(1)}
\usage{
update_phi_AR1(X_list, Y_list, mu)
}
\arguments{
\item{X_list}{List of length N; each \code{X_list[[i]]} is (n-p x d) design matrix X_i.}

\item{Y_list}{List of length N; each \code{Y_list[[i]]} is length n-p response vector Y_i.}

\item{mu}{N x d matrix of node parameters.}
}
\value{
An N x d matrix with the updated phi.
}
\description{
ADMM update for phi of AR(1).
This function will not be export in the final version.
}
