% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_theta}
\alias{update_theta}
\title{Update theta}
\usage{
update_theta(phi, nu, theta, edge_list)
}
\arguments{
\item{phi}{N x p matrix of node parameters (one row per node).}

\item{nu}{E x d matrix of auxiliary variables.}

\item{theta}{E x p matrix of scaled dual variables.}

\item{edge_list}{E x 3 matrix. Each row e is (i, j, w_ij)}
}
\value{
Updated nu matrix of size E x p.
}
\description{
ADMM update for theta
This function will not be export in the final version.
}
