% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_theta_AR1}
\alias{update_theta_AR1}
\title{Update theta of AR(1)}
\usage{
update_theta_AR1(mu, nu, theta, edge_list)
}
\arguments{
\item{nu}{E vector of auxiliary variables.}

\item{theta}{E vector of scaled dual variables.}

\item{edge_list}{E by 3 matrix. Each row e is (i, j, w_ij)}

\item{phi}{N vector of node parameters.}
}
\value{
Updated nu matrix of size E by p.
}
\description{
ADMM update for theta
This function will not be export in the final version.
}
