% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{gen_weighted_block_graph}
\alias{gen_weighted_block_graph}
\title{Generate weighted block graph with 3 or more clusters}
\usage{
gen_weighted_block_graph(
  cluster_sizes = c(50, 50, 50),
  p_in = 0.3,
  p_out = 0.05,
  weight_in_range = c(1, 1.5),
  weight_out_range = c(0.1, 0.5),
  seed = NULL
)
}
\description{
Generate weighted block graph with 3 or more clusters
}
