% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_graph_info}
\alias{get_graph_info}
\title{Get graph information}
\usage{
get_graph_info(adj_w)
}
\arguments{
\item{adj_w}{Numeric matrix (N x N) weighted adjacency matrix. Must be symmetric.}
}
\value{
A list with:
\itemize{
\item edge_list: numeric matrix (i, j, w_ij)
\item node_degree: integer vector of node degrees
}
}
\description{
Convert weighted adjacency matrix to (1) edge list and (2) node degrees
}
\examples{
# 5-node weighted, symmetric adjacency matrix
adj_w <- matrix(0, nrow = 5, ncol = 5)
adj_w[1, 2] <- 1.5; adj_w[1, 4] <- 2.0; adj_w[2, 3] <- 0.8; adj_w[4, 5] <- 3.2
adj_w <- adj_w + t(adj_w)
get_graph_info(adj_w)
}
